% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monk_proc_v2p1.R
\name{monk_proc_v2p1}
\alias{monk_proc_v2p1}
\title{DMN methylation preprocessing
Epi 450K Data Pre-processing is first established on June 10, 2021 By Pamela Scorza
Version 2 monk preprocessing Last update: 2026 09 02: In the process of revision,
preprocessing pipeline updated is required. We updated as following:}
\usage{
monk_proc_v2p1(
  WB,
  idatpath = NULL,
  targetfile = NULL,
  probthresh = 0.01,
  bead_min_count = 3,
  fail_fraction = 0.05,
  snp_maf_dropLoci = 0,
  crosshyb_maf = 0.05,
  rmXY = FALSE,
  sample_fail_fraction = 0.05,
  median_intensity_cutoff = 10.5,
  outfilename = "monkproc_v2p1"
)
}
\arguments{
\item{WB}{methylation file imported by minfi package}

\item{idatpath}{idat file path}

\item{targetfile}{sample information file. This must include Basename field that include idat_id}

\item{probthresh}{0.01 (default) p-value threshold to filter out significantly methylation CPG sites.}

\item{bead_min_count}{3 (default)}

\item{fail_fraction}{0.05 (default)}

\item{snp_maf_dropLoci}{0 (default)}

\item{crosshyb_maf}{0.05 (default)}

\item{rmXY}{FALSE (default) remove XY chromosomes}

\item{sample_fail_fraction}{0.05 (default) In probe-level filtering, probes where qc failire happenes greater than sample_fail_fraction,
will be removed}

\item{median_intensity_cutoff}{10.5 (default) for removing sample.}

\item{outfilename}{monkproc_v2 (devault) output file name}
}
\value{
The resulting beta
}
\description{
1. adding a bead count threshold, and for both detP and bead count,
removing probes that fail either in >5% of samples. (in previous version, we excluded )
}
\details{
2. add Genomic-based probe filtering
}
\examples{
#library(monklab.methyl)
#library(dplyr)
#idatPath <- "/Volumes/ALSPAC/EPI/EPI_methylation_rawdata_froam_ben_2019_12_20/idat" # path of the folder
#targets <- readxl::read_excel("/Volumes/ALSPAC/EPI/EPI_methylation_rawdata_froam_ben_2019_12_20/monk_sample_450K_placenta.xlsx")
#targetfile=targets[1:5,c(1,2,3,4,11,12,13,14,15,17)]
#targetfile$Basename <- targetfile$idat_id#paste0(targets$sentrixbarcode, "_", targets$samplesection) # name of the files
#WB <- read.metharray.exp(base=idatPath, targets=targetfile, verbose=T, extended=TRUE) # read the idat file one by one
#monk_proc_v2(idatpath,targetfile,probthresh=0.01,outfilename='monkproc_v2_test')
}
