% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{Interactive QC Codes}
\usage{
qc(
  WB = NULL,
  idatpath = NULL,
  targetfile = NULL,
  quiet = F,
  sexinfo = NULL,
  minfiqc = TRUE,
  ewastoolqc = TRUE
)
}
\arguments{
\item{quiet}{F (default) do not print data import information.}

\item{sexinfo}{NULL (if specified, we can compare the )}

\item{minfiqc}{TRUE (default) run minfiqc}

\item{ewastoolqc}{TRUE (default) use ewastools to retrieve 17 contraol measures}
}
\value{
qc index
}
\description{
Interactive QC Codes
}
\examples{
Here we see that two samples failed: 9285451058_R01C01 and 9285451058_R06C02
}
