% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monk_proc_v1.R
\name{monk_proc_v1}
\alias{monk_proc_v1}
\title{DMN methylation preprocessing
Epi 450K Data Pre-processing is first established on June 10, 2021 By Pamela Scorza
Version 1 monk preprocessing Last update: 2022 04 08}
\usage{
monk_proc_v1(
  WB,
  idatpath = NULL,
  targetfile = NULL,
  probthresh = 0.01,
  outfilename = "monkproc_v1"
)
}
\arguments{
\item{WB}{methylation file imported by minfi package}

\item{idatpath}{idat file path}

\item{targetfile}{sample information file. This must include Basename field that include idat_id}

\item{probthresh}{0.01 (default) p-value threshold to filter out significantly methylation CPG sites.}

\item{outfilename}{monkproc_v1 (devault) output file name}
}
\value{
The resulting beta
}
\description{
DMN methylation preprocessing
Epi 450K Data Pre-processing is first established on June 10, 2021 By Pamela Scorza
Version 1 monk preprocessing Last update: 2022 04 08
}
\examples{
#library(monklab.methyl)
#library(dplyr)
#idatPath <- "/Volumes/ALSPAC/EPI/EPI_methylation_rawdata_froam_ben_2019_12_20/idat" # path of the folder
#targets <- readxl::read_excel("/Volumes/ALSPAC/EPI/EPI_methylation_rawdata_froam_ben_2019_12_20/monk_sample_450K_placenta.xlsx")
#targetfile=targets[1:5,c(1,2,3,4,11,12,13,14,15,17)]
#targetfile$Basename <- targetfile$idat_id#paste0(targets$sentrixbarcode, "_", targets$samplesection) # name of the files
#WB <- read.metharray.exp(base=idatPath, targets=targetfile, verbose=T) # read the idat file one by one
#monk_proc_v1(idatpath,targetfile,probthresh=0.01,outfilename='monkproc_v1_test')
Distribution of beta-values: before and after noob normalization
notice the blue density traces (raw) are more spread out; background correction brings them together

Barplot of the mean detection p-value by sample
This tells us something (??) about the sample
notice that the type I and II peaks are more closely aligned after rcp adjustment
(particularly in the higher peak)
## Batch effects
This can create batch effects (technical variation) with different intensities by position (row effect).
Other commonly observed batch effects include bisulfite processing plate, chip, and processing date.
We will test 3 potential factors
## Principal Component Analysis (PCA)
Calculate major sources of variability of DNA methylation using PCA
Extract the Principal Components from SVD
ComBat eBayes adjustment using a known variable of interest (here we use row)
PCA after removing batch effects
By Beadchip
}
